update bootstrap
* allow arbitrary certs to be added to bootstrap directory and imported via post hook
* added bash gitdirs alternative if fd is found
* install script refactored to make it easier for tailoring

+++ install.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+CLONE_URL=${CLONE_URL:="https://github.com/vasdee/dotfiles.git"}
+LOCAL_CLONE_PATH=${LOCAL_CLONE_PATH:="private/github.com/vasdee/dotfiles"}
+
 echo "Starting dotfile install"
 
 read -p "Enter your default code base directory. [default: ~/code]" DEFAULT_CODE_DIR
@@ -8,17 +11,25 @@ DEFAULT_CODE_DIR=${DEFAULT_CODE_DIR:-~/code}
 echo "Creating ~/.localsettings file with DEFAULT_CODE_DIR=$DEFAULT_CODE_DIR entry"
 echo "DEFAULT_CODE_DIR=${DEFAULT_CODE_DIR}" >> ~/.localsettings
 
-CLONE_DIR=$DEFAULT_CODE_DIR/private/github.com/vasdee/dotfiles
-echo "Cloning https://github.com/vasdee/dotfiles.git to $CLONEDIR"
+CLONE_DIR=${DEFAULT_CODE_DIR}/${LOCAL_CLONE_PATH}
+echo "Cloning $CLONE_URL to $CLONE_DIR"
 mkdir -p $CLONE_DIR
-git clone https://github.com/vasdee/dotfiles.git $CLONEDIR
-cd $CLONEDIR
+git clone $CLONE_URL $CLONE_DIR
+cd $CLONE_DIR
 
 echo "linking tuckr to ~/.local/bin/tuckr - make sure this is set in your path or run bash or zsh hooks"
-ln -s $PWD/bin/$(uname -s)/$(uname -m)/tuckr $HOME/.local/bin/tuckr
+if [ ! -f $HOME/.local/bin/tuckr ]; then
+    ln -s $PWD/bin/$(uname -s)/$(uname -m)/tuckr $HOME/.local/bin/tuckr
+else
+    echo "link to tuckr in $HOME/.local/bin/tuckr already exists!"
+fi
 
 echo "linking dot files to common config area - ~/.config/dotfiles"
-ln -s $PWD $HOME/.config/dotfiles
+if [ ! -d $HOME/.config/dotfiles ]; then
+    ln -s $PWD $HOME/.config/dotfiles
+else
+    echo "$HOME/.config/dotfiles already exists!"
+fi
 
 LOCAL_BIN="$HOME/.local/bin"
